+++ tests/test_filling_fields.py
@@ -4,6 +4,7 @@ from decimal import Decimal
 from os.path import abspath
 from tempfile import gettempdir
 
+import django
 from django.conf import settings
 from django.core.validators import (
     validate_ipv4_address,
@@ -201,6 +202,15 @@ class TestFillingIntFields:
         small_int_field = models.DummyIntModel._meta.get_field("small_int_field")
         assert isinstance(small_int_field, fields.SmallIntegerField)
         assert isinstance(dummy_int_model.small_int_field, int)
+
+    @pytest.mark.skipif(
+        django.VERSION < (5, 0),
+        reason="The db_default field attribute was added after 5.0",
+    )
+    def test_respects_db_default(self):
+        person = baker.make(models.Person, age=10)
+        assert person.age == 10
+        assert person.retirement_age == 20
 
 
 @pytest.mark.django_db